+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+ - Cresnet Occupancy Sensor Adapters - GLS-ODT-C-CN, GLS-OIR-C-CN, and GLS-OIRLCL-C-CN
 
 ## [6.0.2] - 2019-01-16
 ### Changed